implementation that passes 3rd test

+++ SimilarityCheckKATA/SimilarityCheck.cpp
@@ -9,6 +9,6 @@ public:
 		auto length2 = static_cast<int>(str2.size());
 		auto gap = std::abs(length1 - length2);
 		auto shorter_one = std::min(length1, length2);
-		return (1 - (gap/shorter_one)) * 60;
+		return (1 - (static_cast<double>(gap)/static_cast<double>(shorter_one))) * 60;
 	}
 };
